pyproject.toml
[project]
name = "bob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "networkx>=3.5",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "rich>=13.0",
]
